requirements.txt
# 核心依赖
pandas>=1.3.5
networkx>=2.6.3
pyvis>=0.2.1
nltk>=3.6.7
matplotlib>=3.5.1
seaborn>=0.11.2
pyyaml>=6.0
wordcloud>=1.8.1
requests>=2.28.1
urllib3>=1.26.12

# 向量编码相关依赖
sentence-transformers>=2.2.2
scikit-learn>=1.0.2
torch>=1.13.1
numpy>=1.22.0 